backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
supabase==2.4.2
groq==0.9.0
sentence-transformers==2.7.0
chromadb==0.5.0
PyMuPDF==1.24.3
easyocr==1.7.1
tiktoken==0.7.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
sse-starlette==2.1.0
